Move to .NET 8, remove AutoMapper, fix EF bugs
Newer version of EF Core disabled automatic client side processing for queries that could not be translated to the server-side.
Also had a bug with the EF change tracker freaking out when updating an Effect to have a different statistic value. Solved by changing to explicit load of character and effect statistics.